requirements.txt
# Telegram Bot
python-telegram-bot[webhooks]==20.7

# Supabase
supabase==2.3.4

# HTTP client
httpx==0.27.0

# Environment & config
python-dotenv==1.0.1
pydantic==2.6.0
pydantic-settings==2.2.0

# Data processing
pandas==2.2.0

# Image processing
Pillow==10.2.0

# Dashboard (Stage 3)
streamlit==1.31.0
plotly==5.18.0

# Utilities
python-dateutil==2.8.2
pytz==2024.1
aiofiles==23.2.1
